+++ HexQuickStackStorage/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.4.0
+
+### Added
+- ServerSync integration
+- Configuration option to auto stack trophies
+
 ## v1.3.0
 
 ### Added